+++ internal/provider/resource_monitor_http.go
@@ -135,6 +135,7 @@ func buildHTTPMonitor(ctx context.Context, data *MonitorHTTPResourceModel, diags
 			OAuthClientID:       data.OAuthClientID.ValueString(),
 			OAuthClientSecret:   data.OAuthClientSecret.ValueString(),
 			OAuthScopes:         data.OAuthScopes.ValueString(),
+			OAuthAudience:       data.OAuthAudience.ValueString(),
 			CacheBust:           data.CacheBust.ValueBool(),
 		},
 	}
@@ -236,6 +237,7 @@ func populateHTTPMonitorBaseFieldsForHTTP(httpMonitor *monitor.HTTP, m *MonitorH
 	m.OAuthClientID = stringOrNull(httpMonitor.OAuthClientID)
 	m.OAuthClientSecret = stringOrNull(httpMonitor.OAuthClientSecret)
 	m.OAuthScopes = stringOrNull(httpMonitor.OAuthScopes)
+	m.OAuthAudience = stringOrNull(httpMonitor.OAuthAudience)
 	m.CacheBust = types.BoolValue(httpMonitor.CacheBust)
 }