+++ internal/provider/resource_monitor_mysql.go
@@ -37,6 +37,7 @@ type MonitorMySQLResourceModel struct {
 
 	DatabaseConnectionString types.String `tfsdk:"database_connection_string"`
 	DatabaseQuery            types.String `tfsdk:"database_query"`
+	Conditions               types.List   `tfsdk:"conditions"`
 }
 
 // Metadata returns the metadata for the resource.
@@ -68,6 +69,7 @@ func (*MonitorMySQLResource) Schema(
 				Computed:            true,
 				Default:             stringdefault.StaticString("SELECT 1"),
 			},
+			"conditions": conditionsAttribute(),
 		}),
 	}
 }
@@ -109,6 +111,7 @@ func (r *MonitorMySQLResource) Create(
 		MySQLDetails: monitor.MySQLDetails{
 			DatabaseConnectionString: data.DatabaseConnectionString.ValueString(),
 			DatabaseQuery:            &databaseQuery,
+			Conditions:               buildConditions(ctx, data.Conditions, &resp.Diagnostics),
 		},
 	}
 
@@ -229,6 +232,8 @@ func (r *MonitorMySQLResource) Read(ctx context.Context, req resource.ReadReques
 		data.NotificationIDs = types.ListNull(types.Int64Type)
 	}
 
+	data.Conditions = populateConditions(ctx, mysqlMonitor.Conditions, &resp.Diagnostics)
+
 	data.Tags = handleMonitorTagsRead(ctx, mysqlMonitor.Tags, data.Tags, &resp.Diagnostics)
 	if resp.Diagnostics.HasError() {
 		return
@@ -274,6 +279,7 @@ func (r *MonitorMySQLResource) Update(
 		MySQLDetails: monitor.MySQLDetails{
 			DatabaseConnectionString: data.DatabaseConnectionString.ValueString(),
 			DatabaseQuery:            &databaseQuery,
+			Conditions:               buildConditions(ctx, data.Conditions, &resp.Diagnostics),
 		},
 	}