+++ internal/provider/resource_monitor_postgres.go
@@ -37,6 +37,7 @@ type MonitorPostgresResourceModel struct {
 
 	DatabaseConnectionString types.String `tfsdk:"database_connection_string"`
 	DatabaseQuery            types.String `tfsdk:"database_query"`
+	Conditions               types.List   `tfsdk:"conditions"`
 }
 
 // Metadata returns the metadata for the resource.
@@ -68,6 +69,7 @@ func (*MonitorPostgresResource) Schema(
 				Computed:            true,
 				Default:             stringdefault.StaticString("SELECT 1"),
 			},
+			"conditions": conditionsAttribute(),
 		}),
 	}
 }
@@ -107,7 +109,8 @@ func (r *MonitorPostgresResource) Create(
 		},
 		PostgresDetails: monitor.PostgresDetails{
 			DatabaseConnectionString: data.DatabaseConnectionString.ValueString(),
-			DatabaseQuery:            data.DatabaseQuery.ValueString(),
+			DatabaseQuery:            data.DatabaseQuery.ValueStringPointer(),
+			Conditions:               buildConditions(ctx, data.Conditions, &resp.Diagnostics),
 		},
 	}
 
@@ -204,7 +207,11 @@ func (r *MonitorPostgresResource) Read(ctx context.Context, req resource.ReadReq
 	data.UpsideDown = types.BoolValue(postgresMonitor.UpsideDown)
 	data.Active = types.BoolValue(postgresMonitor.IsActive)
 	data.DatabaseConnectionString = types.StringValue(postgresMonitor.DatabaseConnectionString)
-	data.DatabaseQuery = types.StringValue(postgresMonitor.DatabaseQuery)
+	if postgresMonitor.DatabaseQuery != nil {
+		data.DatabaseQuery = types.StringValue(*postgresMonitor.DatabaseQuery)
+	} else {
+		data.DatabaseQuery = types.StringValue("SELECT 1")
+	}
 
 	if postgresMonitor.Parent != nil {
 		data.Parent = types.Int64Value(*postgresMonitor.Parent)
@@ -223,6 +230,8 @@ func (r *MonitorPostgresResource) Read(ctx context.Context, req resource.ReadReq
 	} else {
 		data.NotificationIDs = types.ListNull(types.Int64Type)
 	}
+
+	data.Conditions = populateConditions(ctx, postgresMonitor.Conditions, &resp.Diagnostics)
 
 	data.Tags = handleMonitorTagsRead(ctx, postgresMonitor.Tags, data.Tags, &resp.Diagnostics)
 	if resp.Diagnostics.HasError() {
@@ -267,7 +276,8 @@ func (r *MonitorPostgresResource) Update(
 		},
 		PostgresDetails: monitor.PostgresDetails{
 			DatabaseConnectionString: data.DatabaseConnectionString.ValueString(),
-			DatabaseQuery:            data.DatabaseQuery.ValueString(),
+			DatabaseQuery:            data.DatabaseQuery.ValueStringPointer(),
+			Conditions:               buildConditions(ctx, data.Conditions, &resp.Diagnostics),
 		},
 	}